Migrated functionality from Settings screen to My Profile screen. Removed the unused or highly faulty functionalities. Made My Profile scrollable.
